crawler/requirements.txt
yt-dlp>=2024.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
pykrx>=1.0.0
notebooklm-py>=0.1.0
fastapi>=0.115.0
uvicorn>=0.34.0
telethon>=1.36.0
